fix: render the conversation in a new tab
